requirements.txt
Django==1.6
argparse==1.2.1
distribute==0.6.34
django-tastypie==0.10.0
django-tastypie-mongoengine==0.4.5
mimeparse==0.1.3
mongoengine==0.8.1
pymongo==2.5
python-dateutil==2.2
six==1.8.0
wsgiref==0.1.2
